Moved the delete view classproperty to DeleteBreadcrumbMixin.
