+++ locale/en_US/LC_MESSAGES/django.po
@@ -8,7 +8,7 @@ msgid ""
 msgstr ""
 "Project-Id-Version: PACKAGE VERSION\n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2021-04-03 17:38-0400\n"
+"POT-Creation-Date: 2021-04-06 09:58-0400\n"
 "PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\n"
 "Last-Translator: FULL NAME <EMAIL@ADDRESS>\n"
 "Language-Team: LANGUAGE <LL@li.org>\n"
@@ -77,48 +77,48 @@ msgstr ""
 msgid "Create"
 msgstr ""
 
-#: demo/views.py:37
+#: demo/views.py:38
 msgid "My Test Breadcrumb"
 msgstr ""
 
-#: demo/views.py:42
+#: demo/views.py:43
 msgid "List tests"
 msgstr ""
 
-#: demo/views.py:77
+#: demo/views.py:78
 msgid "My new home"
 msgstr ""
 
-#: view_breadcrumbs/generic/base.py:29
+#: view_breadcrumbs/generic/base.py:34
 msgid "Home"
 msgstr ""
 
-#: view_breadcrumbs/generic/base.py:35
+#: view_breadcrumbs/generic/base.py:40
 #, python-format
 msgid "%(class_name)s should have a crumbs property."
 msgstr ""
 
-#: view_breadcrumbs/generic/base.py:49
+#: view_breadcrumbs/generic/base.py:54
 msgid "Breadcrumb requires a tuple of label and view name."
 msgstr ""
 
-#: view_breadcrumbs/generic/base.py:69
+#: view_breadcrumbs/generic/base.py:74
 msgid "list"
 msgstr ""
 
-#: view_breadcrumbs/generic/base.py:70
+#: view_breadcrumbs/generic/base.py:75
 msgid "create"
 msgstr ""
 
-#: view_breadcrumbs/generic/base.py:71
+#: view_breadcrumbs/generic/base.py:76
 msgid "update"
 msgstr ""
 
-#: view_breadcrumbs/generic/base.py:72
+#: view_breadcrumbs/generic/base.py:77
 msgid "delete"
 msgstr ""
 
-#: view_breadcrumbs/generic/base.py:73
+#: view_breadcrumbs/generic/base.py:78
 msgid "detail"
 msgstr ""
 
@@ -145,7 +145,7 @@ msgstr ""
 msgid "Update: %(instance)s"
 msgstr ""
 
-#: view_breadcrumbs/utils.py:44
+#: view_breadcrumbs/utils.py:45
 #, python-format
 msgid "%(model)s is not installed or missing from the app registry."
 msgstr ""
