+++ locale/fr/LC_MESSAGES/django.po
@@ -8,7 +8,7 @@ msgid ""
 msgstr ""
 "Project-Id-Version: PACKAGE VERSION\n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2021-04-03 17:38-0400\n"
+"POT-Creation-Date: 2021-04-06 09:58-0400\n"
 "PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\n"
 "Last-Translator: FULL NAME <EMAIL@ADDRESS>\n"
 "Language-Team: LANGUAGE <LL@li.org>\n"
@@ -78,48 +78,48 @@ msgstr "Effacer"
 msgid "Create"
 msgstr "Créer"
 
-#: demo/views.py:37
+#: demo/views.py:38
 msgid "My Test Breadcrumb"
 msgstr "Mon fil d'Ariane de test"
 
-#: demo/views.py:42
+#: demo/views.py:43
 msgid "List tests"
 msgstr "Liste des tests"
 
-#: demo/views.py:77
+#: demo/views.py:78
 msgid "My new home"
 msgstr "Ma nouvelle maison"
 
-#: view_breadcrumbs/generic/base.py:29
+#: view_breadcrumbs/generic/base.py:34
 msgid "Home"
 msgstr "Domicile"
 
-#: view_breadcrumbs/generic/base.py:35
+#: view_breadcrumbs/generic/base.py:40
 #, python-format
 msgid "%(class_name)s should have a crumbs property."
 msgstr "%(class_name)s doit avoir une propriété crumbs."
 
-#: view_breadcrumbs/generic/base.py:49
+#: view_breadcrumbs/generic/base.py:54
 msgid "Breadcrumb requires a tuple of label and view name."
 msgstr "Le fil d'Ariane nécessite un tuple d'étiquette et de nom de vue."
 
-#: view_breadcrumbs/generic/base.py:69
+#: view_breadcrumbs/generic/base.py:74
 msgid "list"
 msgstr "liste"
 
-#: view_breadcrumbs/generic/base.py:70
+#: view_breadcrumbs/generic/base.py:75
 msgid "create"
 msgstr "créer"
 
-#: view_breadcrumbs/generic/base.py:71
+#: view_breadcrumbs/generic/base.py:76
 msgid "update"
 msgstr "mettre à jour"
 
-#: view_breadcrumbs/generic/base.py:72
+#: view_breadcrumbs/generic/base.py:77
 msgid "delete"
 msgstr "effacer"
 
-#: view_breadcrumbs/generic/base.py:73
+#: view_breadcrumbs/generic/base.py:78
 msgid "detail"
 msgstr "détail"
 
@@ -146,7 +146,7 @@ msgstr ""
 msgid "Update: %(instance)s"
 msgstr "Mise à jour: %(instance)s"
 
-#: view_breadcrumbs/utils.py:44
+#: view_breadcrumbs/utils.py:45
 #, python-format
 msgid "%(model)s is not installed or missing from the app registry."
 msgstr "%(model)s n'est pas installé ou absent du registre de l'application."
