+++ view_breadcrumbs/generic/__init__.py
@@ -3,3 +3,4 @@ from .create import CreateBreadcrumbMixin  # noqa
 from .detail import DetailBreadcrumbMixin  # noqa
 from .list import ListBreadcrumbMixin  # noqa
 from .update import UpdateBreadcrumbMixin  # noqa
+from .delete import DeleteBreadcrumbMixin  # noqa
