+++ view_breadcrumbs/generic/base.py
@@ -84,14 +84,3 @@ class BaseModelBreadcrumbMixin(BaseBreadcrumbMixin):
     @property
     def model_name_title_plural(self):
         return get_verbose_name_plural(self.model).title()
-
-    @classproperty
-    def delete_view_name(self):
-        return action_view_name(self.model, self.delete_view_suffix, full=False)
-
-    @property
-    def __delete_view_name(self):
-        return action_view_name(self.model, self.detail_view_suffix)
-
-    def delete_view_url(self, instance):
-        return reverse(self.__delete_view_name, kwargs={"pk": instance.pk})
